@@ -1,5 +1,6 @@
 package se.andreas.engagementapp.account.dto;
 
+import jakarta.json.bind.annotation.JsonbCreator;
 import jakarta.json.bind.annotation.JsonbVisibility;
 import se.andreas.engagementapp.account.model.FieldVisibilityStrategy;
 
@@ -7,20 +8,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 @JsonbVisibility(FieldVisibilityStrategy.class)
-public final class Account {
+public record Account(String name, List<BusinessObject> businessObjects) {
 
-    private String name;
-    private List<BusinessObject> businessObjects;
-
-    public Account() {
-        this.businessObjects = new ArrayList<>();
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public List<BusinessObject> getBusinessObjects() {
-        return businessObjects;
+    @JsonbCreator
+    public Account {
     }
 }
