@@ -1,26 +1,15 @@
 package se.andreas.engagementapp.account.dto;
 
+import jakarta.json.bind.annotation.JsonbCreator;
 import jakarta.json.bind.annotation.JsonbVisibility;
 import se.andreas.engagementapp.account.model.FieldVisibilityStrategy;
 
 import java.math.BigDecimal;
 
 @JsonbVisibility(FieldVisibilityStrategy.class)
-public final class BalanceInformation implements BusinessObject {
+public record BalanceInformation(BigDecimal balance, String currency, BigDecimal interest) implements BusinessObject {
 
-    private BigDecimal balance;
-    private String currency;
-    private BigDecimal interest;
-
-    public BigDecimal getBalance() {
-        return balance;
-    }
-
-    public String getCurrency() {
-        return currency;
-    }
-
-    public BigDecimal getInterest() {
-        return interest;
+    @JsonbCreator
+    public BalanceInformation {
     }
 }
