@@ -3,15 +3,7 @@ package se.andreas.engagementapp.account.model;
 import jakarta.json.bind.annotation.JsonbVisibility;
 
 @JsonbVisibility(FieldVisibilityStrategy.class)
-public class AccountRole {
-
-    private final String name;
-    private final Role role;
-
-    public AccountRole(String name, Role role) {
-        this.name = name;
-        this.role = role;
-    }
+public record AccountRole(String name, Role role) {
 
     public enum Role {
         OWNER, OTHER;
